feat: Add predicate for flyspell to enable spell checking
When `flyspell-mode` (or `flyspell-prog-mode`) is turned on in a buffer that's
major mode is `git-commit-ts-mode`, the spell checking doesn't work. This
commit alleviates that by adding a predicate that.  specifies which words
should be spell checked. The predicate is registered as a
`flyspell-mode-predicate` property of `git-commit-ts-mode`.

I think that's it's most important to spell check subject and message body
including trailers and breaking changes. However, it's probably best not to
spell check types, scopes, nor tokens of trailers. Also removed email addresses
and GitHub style mentions (things that starts with @) from words to check.